# engine/api/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "media_get_output": {
        "additionalProperties": false,
        "properties": {
          "media": {
            "additionalProperties": false,
            "properties": {
              "checksum": {
                "type": "string"
              },
              "display_path": {
                "type": [
                  "null",
                  "string"
                ]
              },
              "error": {
                "type": [
                  "null",
                  "string"
                ]
              },
              "height": {
                "type": [
                  "null",
                  "integer"
                ]
              },
              "id": {
                "type": "string"
              },
              "mime_type": {
                "type": "string"
              },
              "size_bytes": {
                "type": "integer"
              },
              "state": {
                "type": "string"
              },
              "thumbnail_path": {
                "type": [
                  "null",
                  "string"
                ]
              },
              "upload_path": {
                "type": "string"
              },
              "version": {
                "type": "integer"
              },
              "width": {
                "type": [
                  "null",
                  "integer"
                ]
              }
            },
            "required": [
              "id",
              "version",
              "checksum",
              "size_bytes",
              "mime_type",
              "state",
              "width",
              "height",
              "error",
              "upload_path",
              "display_path",
              "thumbnail_path"
            ],
            "type": "object"
          }
        },
        "required": [
          "media"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "routerBearer": {
        "bearerFormat": "JWT",
        "description": "Verified by the identity router; the private engine has no authentication or user identity handling.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Wardrobe inventory, dated outfits and private photographs. One registry serves HTTP and MCP.",
    "title": "Retro Wardrobe Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/media/{id}/content": {
      "get": {
        "description": "Read a ready JPEG derivative; original sources are never served.",
        "operationId": "media_content",
        "parameters": [
          {
            "in": "query",
            "name": "variant",
            "required": true,
            "schema": {
              "enum": [
                "display",
                "thumbnail"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Safe photo derivative"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid variant"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authentication required by router"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Photo not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Photo not ready"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Photo storage unavailable"
          }
        },
        "security": [
          {
            "routerBearer": []
          }
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "put": {
        "description": "Upload the exact reserved JPEG/PNG bytes, maximum 12 MiB. Identical retries are safe.",
        "operationId": "media_upload",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            },
            "image/jpeg": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            },
            "image/png": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/media_get_output"
                }
              }
            },
            "description": "Photo status"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid bytes, size or checksum"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Authentication required by router"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Photo not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Photo exceeds 12 MiB"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Photo storage unavailable"
          }
        },
        "security": [
          {
            "routerBearer": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Trusted identity router; direct engine uses /",
      "url": "/retro"
    }
  ]
}